[Objc] - Common protocol for Api and added NSParameterAssert if missing required param + calling completion block with error rather than throwing an exception

+++ samples/client/petstore/objc/SwaggerClient/Core/SWGApiClient.h
@@ -60,7 +60,7 @@ extern NSString *const SWGResponseObjectErrorKey;
  *
  * @return The size of `queuedRequests` static variable.
  */
-+(unsigned long)requestQueueSize;
++(NSUInteger)requestQueueSize;
 
 /**
  * Sets the client unreachable
